requirements.txt
# 深度学习框架
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.30.0

# 数据处理
numpy>=1.24.0
pandas>=2.0.0
opencv-python>=4.8.0
Pillow>=10.0.0

# 自然语言处理
nltk>=3.8.0
spacy>=3.6.0

# 科学计算
scipy>=1.11.0
scikit-learn>=1.3.0

# 可视化
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# 工具库
tqdm>=4.65.0
tensorboard>=2.13.0
wandb>=0.15.0

# 配置文件
pyyaml>=6.0
omegaconf>=2.3.0

# Jupyter
jupyter>=1.0.0
ipywidgets>=8.0.0

# 其他
requests>=2.31.0
tqdm>=4.65.0
